Node weights and property (#100)
* Rename Weights interface to Relationship Weights
* Rename property methods to relationshipWeight in GraphSetup und loaders
* Add possibility to specify node weights
* Add possibility to load some extra property
* Add new methods to weight mapping

- get with default value in the method: to be used when the default value depends on some runtime value
- get with int as key (default impl) to be used when one wants to implement weights for nodes

* Have HeavyGraph implement all latest weight mapping interfaces

+++ core/src/main/java/org/neo4j/graphalgo/core/leightweight/LightGraph.java
@@ -214,7 +214,7 @@ public class LightGraph implements Graph {
             int offset = cursor.offset;
             final int limit = cursor.length + offset;
             while (offset < limit) {
-                consumer.accept(node, array[offset], relMap.get(offset), weightMap.get(offset++));
+                consumer.accept(node, array[offset], relMap.get(offset), weightMap.get((long) offset++));
             }
         }
     }
